First pass at a Line component
Gotta document this design, but the tl;dr is:

Marks (like line) have
  Encodings (like x and y) have
    Scales (like Linear or Divergent)
      Have domains and ranges

The optional magic trick is that since scales are class instances now,
we can qualify scales dynamically based on a data set. So the Line
component can have an xScale with an unqualified domain, but if the x
encoding is a field encoding then we can compute the extent within Line
just in time.