Gridlines component
This can be seen as a derivative of the Axis component, and this pattern
is typically implementing just using an axis in vanilla d3, but I think
it's common enough to justify its own component. No more clever negative
dimensions or removing text you didn't want added in the first place.