Super unfortunate tooling thing
Sometimes cached isn't available in the @glimmer/tracking package, which
is a bit of a phantom package that is supposed to get babelified as part
of ember build by way of ember-auto-import.

Since this bug is upstream, this is used to at least preserve
compatability even if it means @cached isn't being used in old Ember.