Extend Lineal::Points to allow for yielding the computed dataset
This follows the pattern seen in Lineal::Arcs where processing the data
is desired, but rendering it may not be.

The one difference is it's common to want to render points _and other
stuff_ such as text labels. So to prevent needing to invocations of
Lineal::Points, the component has a @renderCircles arg to do double
duty while yielding.