Support automatically shrinking rounded rect radii
When radii sum to more than the width/height of the rectangle, normally
this creates inverted geometry. However, now when safe = true, the radii
will be shrunk proportionally to sum to the width/height, so geometry
never inverts and no one ever has to twiddle this arithmetic on their
own.